# perf: Add object rvalue overload for accessors. Enables reference stealing