--- a/include/pybind11/pybind11.h
+++ b/include/pybind11/pybind11.h
@@ -2057,17 +2057,18 @@ struct enum_base {
             [](const object &arg) { return int_(arg); }, name("__hash__"), is_method(m_base));
     }
 
-    PYBIND11_NOINLINE void value(char const *name_, object value, const char *doc = nullptr) {
+    PYBIND11_NOINLINE void
+    value(char const *name_, const object &value, const char *doc = nullptr) {
         dict entries = m_base.attr("__entries");
         str name(name_);
         if (entries.contains(name)) {
             std::string type_name = (std::string) str(m_base.attr("__name__"));
-            throw value_error(type_name + ": element \"" + std::string(name_)
+            throw value_error(std::move(type_name) + ": element \"" + std::string(name_)
                               + "\" already exists!");
         }
 
         entries[name] = std::make_pair(value, doc);
-        m_base.attr(name) = value;
+        m_base.attr(std::move(name)) = value;
     }
 
     PYBIND11_NOINLINE void export_values() {