--- a/include/pybind11/pytypes.h
+++ b/include/pybind11/pytypes.h
@@ -689,8 +689,8 @@ public:
     }
 
 private:
-    object ensure_object(object &&o) { return std::move(o); }
-    object ensure_object(handle h) { return reinterpret_borrow<object>(h); }
+    static object ensure_object(object &&o) { return std::move(o); }
+    static object ensure_object(handle h) { return reinterpret_borrow<object>(h); }
 
     object &get_cache() const {
         if (!cache) {