--- a/tests/test_pytypes.cpp
+++ b/tests/test_pytypes.cpp
@@ -7,7 +7,6 @@
     BSD-style license that can be found in the LICENSE file.
 */
 
-#include "pybind11/detail/common.h"
 #include "pybind11_tests.h"
 
 #include <utility>