--- a/tests/test_pytypes.py
+++ b/tests/test_pytypes.py
@@ -710,3 +710,9 @@ def test_list_rvalue_getter():
     pop = 1000
     my_list = list(range(pop))
     m.list_rvalue_getter(my_list)
+
+
+def test_populate_dict_rvalue():
+    pop = 1000
+    my_dict = {i: i for i in range(pop)}
+    assert m.populate_dict_rvalue(pop) == my_dict