@rwgk 2022-05-25
I tend to be deliberately a bit ignorant towards low-level mechanics***, but trying to understand this PR, is the following correct?

* Both `handle` and `object` are just `PyObject *` underneath.
* Copying a pointer or moving a pointer does not make a difference in terms of runtime performance.
* Deleting an `object` means DECREF, and creating a new one with `reinterpret_borrow<object>` means `INCREF`.

Are those DECREF/INCREF what this PR is eliminating?

If that's the point, what kind of test could ensure that this optimization is not accidentally undone?

Also, did you already run a micro benchmark to quantify the best-case performance gain? — Something temporary and quick & dirty to get a rough idea would seem sufficient.

________________________________________________________________

*** I'm generally more focused on high-level aspects & safety than low-level efficiency considerations, trusting that modern compilers will optimize out obvious inefficiencies. I believe cluttering or complicating code to squeeze out a tiny bit of extra efficiency negatively impacts readability & long-term maintainability, and it makes future development work & refactoring more bug-prone. That could result in a significant waste of human time, the most expensive resource we have, just to save a tiny amount of very cheap machine time.