@Skylion007 2022-05-25
>> Are those DECREF/INCREF what this PR is eliminating?
Yes