@rwgk 2022-05-25
> > > Are those DECREF/INCREF what this PR is eliminating?
> > > Yes

What test could ensure that this optimization is not accidentally undone?