@Skylion007 2022-05-25
>>> What test could ensure that this optimization is not accidentally undone?
Writing the tests for that would be pretty difficult. I would normally use constructor_stats.h, but we would need to override the ctors and dtor of the object class to measure that. Any thoughts on how to do it @rwgk? I think it's more of a style issue TBH. reinterpret_borrow should not be used on objects. Maybe we could add an overload to reinterpret_borrow that throws an exception if it's an pyobject (to prevent the implicit conversion of a handle). There is no reason to use it with objects, but it can be done due to the implicit handle conversion. Thoughts @rwgk?