@Skylion007 2022-05-25
@rwgk TLDR: the real evil here is the antipattern of `objects -> handle -> objects` it does incref / decrefs for no reason. I am not sure of a way to guard against it though since `handle -> object` are a valid use case. I normally agree with you about trusting the compiler to do this stuff for us, but this is an obvious case where it is failing, the overhead of maintainig these overloads is negligible.