@Skylion007 2022-05-25
@rwgk Also, I wouldn't normally add these extra overloads just for that, but the fact that is such a hot code path (any attr accesses) made me want to optimize it.