@rwgk 2022-05-26
To me it feels like we're driving blindly (no performance measurements) and without guard rails (test to ensure the optimization is not accidentally undone). Re performance measurements: I believe that is only a few minutes worth of effort, to get a rough idea at least; i.e. we're driving blindly unnecessarily. Re tests, letting basic best practices slip needs to be clearly stated in the PR description, with a rationale, to prove (incl. to ourselves) that we carefully thought about it.