@Skylion007 2022-05-26
@rwgk Care to do the benchmarking then? We currently don't have any benchmarking framework on master at all currently.