@Skylion007 2022-05-26
@rwgk Googling, this is effectively the same as benchmarking incref / decref and I don't expect any performance difference there from the CPython benchmarks. Do not expect any speed significant difference from this PR anyhow. I'll update the PR description with this explanation.