@rwgk 2022-05-26
> @rwgk Care to do the benchmarking then? We currently don't have any benchmarking framework on master at all currently.

We don't need one for this (although it would be nice to have), a quick 10 minute thing like this will do:

https://github.com/rwgk/rwgk_tbx/blob/main/test_perf_error_already_set.cpp
https://github.com/rwgk/rwgk_tbx/blob/main/test_perf_error_already_set.py

Just post a link to your code and the results here.