@rwgk 2022-05-26
I've been staring at this code for a while:
```
    template <typename T>
    void operator=(T &&value) & {
        get_cache() = ensure_object(object_or_cast(std::forward<T>(value)));
    }
```
With
```
private:
    static object ensure_object(object &&o) { return std::move(o); }
    static object ensure_object(handle h) { return reinterpret_borrow<object>(h); }

    object &get_cache() const {
        if (!cache) {
            cache = Policy::get(obj, key);
        }
        return cache;
    }
private:
    handle obj;
    key_type key;
    mutable object cache;```
```

I.e. in `operator=(T &&value)`, on the lhs we have a `object &` (r-value) and on the right side `object` (r-value) IIUC.

Is the compiler forced or only permitted to elide the copy here?

Mainly I was looking for a way to test that the INCREF/DECREF cycle is in fact eliminated, but I failed to see one.

But what stood out to me, does `cache` have to be `object`? Could we make it `handle`? Then we have obvious & full control over when INCREF and DECREF are invoked. AFAICT, if we update the (copy, move) x (ctor, operator=) and dtor accordingly we should be fine. Could that unlock other manual optimizations?

Generally thinking:

* If we had a test that proves the INCREF/DECREF cycle is in fact eliminated, ok. We don't know how much that helps, but at least we're sure it's gone and won't come back.
* If we could at least measure a gain, at least one platform, ok.

But without either I feel very uneasy about making the code more complex.

If none of us can see a way to implement a test, then a rough benchmark would be useful. If and only if that gives us measurements that prove it's worth the trouble, the `handle` approach might be interesting, although with such measurements I'd also be OK with this PR as-is.