@Skylion007 2022-05-26
>> Is the compiler forced or only permitted to elide the copy here? 
It's explicitly allowed to on C++14, but not requited to elide it. It's explicitly required to elide the copy on C++17.