@rwgk 2022-05-26
> It's explicitly allowed to on C++14, but not requited to elide it. It's explicitly required to elide the copy on C++17. I think it should elide it on C++11.

I looked around some more and now I'm thinking that my "Is the compiler forced or only permitted to elide the copy here?" was actually besides the point. The correct question: "Is the compiler forced or only permitted to use the move ctor here?" I don't know all the rules that determine that, but my guess is it is actually forced to use the move ctor here. So that's great.

Could you suggest a code snippet for a best-case micro benchmark? Just the part to loop over, I could plug that into my ad-hoc benchmark code I pointed out earlier. That's really quick.