@Skylion007 2022-05-26
> The correct question: "Is the compiler forced or only permitted to use the move ctor here?" I don't know all the rules that determine that, but my guess is it is actually forced to use the move ctor here. So that's great.

Yes, it's forced to use the move ctor.