@rwgk 2022-05-26
Just to let you know, **in a rush** (I'll be out for a few hours now)

the previous was not hitting the code path we discussed yesterday

by running all unit tests I found exactly one case that hits it

this is what I'm seeing right now

```
master PERF 204265501,4.193s,0.205279μs
pr3970 PERF 216456859,4.168s,0.192548μs
```

```
                 py::object obj,
                 const py::object &attr_name,
                 const py::object &attr) {
    auto l = py::list(1);
    l[0] = 0;
    auto var = l[0];
    while (num_iterations) {
        var = num_iterations;
        num_iterations--;
    }
}
```